first small refactor splitting of contracts and oauth interfaces
