+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/google/uuid v1.1.2
 	github.com/jasonlvhit/gocron v0.0.0-20191007145845-57f89394836a // indirect
 	github.com/labstack/echo/v4 v4.1.17
+	github.com/mdp/qrterminal v1.0.1
 	github.com/mdp/qrterminal/v3 v3.0.0
 	github.com/nuts-foundation/nuts-crypto v0.15.1-0.20201008093701-3776f2936ef8
 	github.com/nuts-foundation/nuts-go-core v0.15.0
